simdutf: add build tag "libsimdutf" to use system simdutf library
This commit also reverts the changes made in bea03ad and moves the
simdutf library files back to the project root (instead of being in
the ./simdutf directory).

NB: For some reason I'm getting spurious build errors on macOS using
the dev toolchain - these errors don't occur when using GCC or when
building on Linux.